ENH: Improve handling of multiple range axes in NRRD reading
In case of images where there are more than one "range" (i.e. non-domain) axes, ITK represents the first one as components in the voxel, but then need to handle the rest as quasi-domain axes, meaning that they need to have a valid (non-NaN) spacing and direction so that the region handling does not fail. This commit adds these missing spacing and direction information for the second and further range axes.